fix: resolve macOS build errors
- spider_io.cpp: fix narrowing conversion from uint64 to int with static_cast
- utility.cpp: update deprecated Boost.Asio API (io_service → io_context, resolver::query → resolve() with strings)

+++ src/cvc/spider_io.cpp
@@ -740,7 +740,7 @@ namespace CVC_NAMESPACE
       using namespace boost;
       thread_info ti(BOOST_CURRENT_FUNCTION);
 
-      int dim[3] = { dimension[0], dimension[1], dimension[2] };
+      int dim[3] = { static_cast<int>(dimension[0]), static_cast<int>(dimension[1]), static_cast<int>(dimension[2]) };
       boost::scoped_array<float> data(new float[dimension.size()]);
 
       if(boundingBox.isNull())
@@ -810,7 +810,7 @@ namespace CVC_NAMESPACE
 	  for(uint64 i = 0; i < wvol.XDim(); i++)
 	    vol(i+off_x,j+off_y,k+off_z, wvol(i,j,k));
 
-      int dim[] = { vol.XDim(), vol.YDim(), vol.ZDim() };
+      int dim[] = { static_cast<int>(vol.XDim()), static_cast<int>(vol.YDim()), static_cast<int>(vol.ZDim()) };
       vol.voxelType(Float); //make sure we have a float volume (though we shouldnt get here if we dont)
       int retval = writeSpiderFile(filename.c_str(),'V',dim,reinterpret_cast<float*>(*vol));
       if(retval)
